Make includable as a uniform CMake dependency
This makes the way PEGTL is included in the project transparent to the
consuming project. If PEGTL is included with `add_subdirectory`
subsequent calls to `find_package` will find the PEGTL version included
with `add_subdirectory`. If PEGTL is included with `find_package`
subsequent calls to `add_subdirectory` will use the PEGTL version
included with find_package.

Since CMake targets are global there exists no way for multiple versions
of PEGTL to exist within one CMake build. Therefore, a FATAL_ERROR is
signalled if an attempt is made to include multiple different versions
of PEGTL.