parse_tree::basic_node now uses template Source, defaults to string_view
There is no reason to copy the string over and over again, most of the
time the source will be alive and a string_view will be enough.

If there is any case where this is not the case, then `basic_node` can
be told to use a different type via the second template parameter.

Closes #243