Add a CMake option to select no std::filesystem
Cmake option PEGTL_NO_STD_FILESYSTEM will mean there is
*no* implementation of std::filesystem selected at compile
time. This will invariably lead to a broken PEGTL if you set
it and use file operations, but it is intended to allow a configuration
on systems without boost or std::filesystem, which don't need
filesystem access but do want memory parsing. This allows
taojson memory parsers to compile back to macOS 10.13, for instance.

See discussion in issue #347